move duplicate and merge previous food blocks functionality into a component
