+++ pages/day/[date]/blocks.js
@@ -5,6 +5,7 @@ import orangeIcon from '../../../public/icons/orange.svg';
 
 import Nav from '../../../shared/components/Nav';
 import Modal from '../../../shared/components/modal';
+import DuplicateAndMergeFoodBlocksFromPreviousDate from '../../../shared/components/DuplicateAndMergePreviousFoodBlocks';
 import Snackbar from '../../../shared/components/Snackbar/Snackbar';
 
 import {
@@ -12,7 +13,6 @@ import {
     updateFoodBlockInLocalStorage,
     removeFoodBlockFromLocalStorage,
     addFoodBlockToLocalStorage,
-    duplicateAndMergeFoodBlocksFromPreviousDate,
     getFoodDictionaryFromLocalStorage
 } from '../../../shared/food/food';
 
@@ -149,28 +149,12 @@ export default function Blocks () {
         if (router.query.modalOpen === 'true') {
             setIsModalOpen(true);
         }
-
-        setAllFoodBlocks(JSON.parse(localStorage.getItem('foodBlocks')));
     }, []);
 
 
     const [ isModalOpen, setIsModalOpen ] = useState(false);
     const toggleModal = () => {
         setIsModalOpen(previousIsModalOpen => !previousIsModalOpen);
-    }
-
-    const [ allFoodBlocks, setAllFoodBlocks ] = useState(null);
-    const [ previousFoodBlockDate, setPreviousFoodBlockDate ] = useState('');
-    const handlePreviousFoodBlockDateChange = event => {
-        if (event.currentTarget.value) return setPreviousFoodBlockDate(event.currentTarget.value);
-    }
-    const handleDuplicateAndMergePreviousDateSubmit = event => {
-        event.preventDefault();
-        if (!previousFoodBlockDate) return;
-
-        duplicateAndMergeFoodBlocksFromPreviousDate({ previousDate: previousFoodBlockDate, currentDate: date });
-        setFoodBlocks(getFoodFromLocalStorage(date));
-        toggleModal();
     }
 
     return (
@@ -368,36 +352,8 @@ export default function Blocks () {
                         add a new food block
                     </button>
 
-                    { allFoodBlocks && Object.keys(allFoodBlocks).length > 1 ? (
-                        <form onSubmit={ handleDuplicateAndMergePreviousDateSubmit }>
-                            <div className="flex flex-direction-column mb-2">
-                                <label
-                                    className="text-black text-medium mb-1"
-                                    htmlFor="foodBlocksDate">duplicate previous day food blocks</label>
-                                <select
-                                    onChange={ handlePreviousFoodBlockDateChange }
-                                    className="text-medium w-50"
-                                    id="foodBlocksDate">
-                                    <option value="">choose a date</option>
-                                    { Object.keys(allFoodBlocks).sort((dateOne, dateTwo) => {
-                                        return new Date(dateOne) - new Date(dateTwo);
-                                    }).slice(-10).map(foodBlockDate => {
-                                        if (foodBlockDate === date) return;
-
-                                        return (
-                                            <option
-                                                key={ foodBlockDate }
-                                                value={ foodBlockDate }>{ foodBlockDate }</option>
-                                        )
-                                    }) }
-                                </select>
-                            </div>
-
-                            <button className="button button-pink">duplicate</button>
-                        </form>
-                    ) : (null) }
-
-                </div> 
+                    <DuplicateAndMergeFoodBlocksFromPreviousDate toggleModal={ toggleModal } setFoodBlocks={ setFoodBlocks } />
+                </div>
             </Modal>
